refactor(rag): share PageIndex KB detection
